refactor(entry): move CLI entry to root main.py; remove src/main.py per new entrypoint policy
